Create interfaces folder and move dataaccess interface to it
