+++ WordleClash.Data/DataAccess.cs
@@ -1,5 +1,5 @@
 using MySql.Data.MySqlClient;
-using WordleClash.Core.DataAccess;
+using WordleClash.Core.Interfaces;
 
 namespace WordleClash.Data;
